requirements.txt
Django==1.8.1
argparse==1.2.1
django-browserid==1.0.0
django-classy-tags==0.6.2
django-missing==0.1.13
django-mongo-auth==0.1.2
django-rest-framework-mongoengine==2.0.2
django-sekizai==0.8.2
djangorestframework==3.1.2
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.9.0
oauthlib==1.0.3
pymongo==2.8
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.4.0
wsgiref==0.1.2
